Revert "With db"
